Funções de controle do index.js movidos para padrao.js
